+++ test/specs/BrowserDetector.test.ts
