+++ test/specs/VarCache.test.ts
